fixup! Call new SOCKS code from parse_socks, to parse multiple packets in row

+++ src/or/proto_socks.c
@@ -671,7 +671,7 @@ handle_socks_message(const uint8_t *raw_data, size_t datalen,
 {
   socks_result_t res = SOCKS_RESULT_DONE;
 
-  uint8_t socks_version = (uint8_t)raw_data[0];
+  uint8_t socks_version = raw_data[0];
 
   if (socks_version == 1)
     socks_version = 5; // SOCKS5 username/pass subnegotiation
